Add more docs and move fixture to setup.py.
